requirements.txt
tensorflow-cpu == 2.3.0
pyqt5 == 5.15.6
pillow == 9.0.1
opencv-python ==4.5.5.62
matplotlib == 3.5.1
scikit-learn==1.0.2
Flask==2.0.2
Flask-Cors==3.0.9